Consolidate stories
